# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task API",
    "version": "1.0",
    "description": "In-memory CRUD API for tasks. Data is held in a plain array and is lost when the server restarts."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/tasks": {
      "get": {
        "summary": "List tasks",
        "operationId": "listTasks",
        "responses": {
          "200": {
            "description": "Matching tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "done filter must be true or false"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "done",
            "in": "query",
            "description": "Filter by completion status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Filter tasks whose title contains this word",
            "schema": {
              "type": "string",
              "example": "memory"
            }
          }
        ]
      },
      "post": {
        "summary": "Create a task",
        "operationId": "createTask",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "title must be a non-empty string"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Task id",
          "schema": {
            "type": "integer",
            "example": 1
          }
        }
      ],
      "get": {
        "summary": "Get one task",
        "operationId": "getTask",
        "responses": {
          "200": {
            "description": "Task found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Task 99 not found"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a task",
        "operationId": "updateTask",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "noFields": {
                    "value": {
                      "error": "provide title or done to update"
                    }
                  },
                  "emptyTitle": {
                    "value": {
                      "error": "title must not be empty"
                    }
                  },
                  "invalidDone": {
                    "value": {
                      "error": "done must be a boolean value"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Task 99 not found"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "operationId": "deleteTask",
        "responses": {
          "204": {
            "description": "Task deleted, no body returned"
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Task 99 not found"
                }
              }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "Task counts",
        "operationId": "getStats",
        "responses": {
          "200": {
            "description": "Aggregate task statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stats"
                }
              }
            }
          }
        }
      }
    },
    "/reset": {
      "post": {
        "summary": "Reset to seed data",
        "operationId": "resetTasks",
        "responses": {
          "200": {
            "description": "Seed tasks restored",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "required": [
          "id",
          "title",
          "done"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Finish DSA assignment"
          },
          "done": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "CreateTaskRequest": {
        "type": "object",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string",
            "example": "Review pointer arithmetic notes"
          }
        }
      },
      "UpdateTaskRequest": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "title": {
            "type": "string",
            "example": "Review pointer arithmetic notes again"
          },
          "done": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "Stats": {
        "type": "object",
        "required": [
          "total",
          "done",
          "open"
        ],
        "properties": {
          "total": {
            "type": "integer",
            "example": 7
          },
          "done": {
            "type": "integer",
            "example": 3
          },
          "open": {
            "type": "integer",
            "example": 4
          }
        }
      }
    }
  }
}